Feature/types and exceptions

+++ src/modules/companies/companies.service.ts
@@ -1,5 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { checkValidUUID } from 'src/common/checkValidUUID';
 import { Repository } from 'typeorm';
 import { Responsible } from '../responsibles/entities/responsible.entity';
 import { CreateCompanyDto } from './dto/create-company.dto';
@@ -20,20 +25,72 @@ export class CompaniesService {
     const savedCompany = await this.companyRepository.save(createdCompany);
 
     if (createCompanyDto.responsibles) {
-      await Promise.all(
-        createCompanyDto.responsibles.map(async (responsible: Responsible) => {
-          const createdResponsible = this.responsibleRepository.create({
-            ...responsible,
-            company: {
-              id: savedCompany.id,
-            },
-          });
-          await this.responsibleRepository.save(createdResponsible);
-        }),
+      await this.handleCreateCompanyResponsibles(
+        savedCompany.id,
+        createCompanyDto.responsibles,
       );
     }
 
     return savedCompany;
+  }
+
+  async handleCreateCompanyResponsibles(
+    companyId: string,
+    responsiblesArray: Responsible[],
+  ) {
+    await Promise.all(
+      responsiblesArray.map(async (responsible: Responsible) => {
+        const createdResponsible = this.responsibleRepository.create({
+          ...responsible,
+          company: {
+            id: companyId,
+          },
+        });
+        await this.responsibleRepository.save(createdResponsible);
+      }),
+    );
+  }
+
+  async handleUpdateCompanyResponsibles(
+    companyId: string,
+    responsiblesArray: Responsible[],
+    previousResponsibles: Responsible[],
+  ) {
+    let responsiblesToRemove: Responsible[] = previousResponsibles;
+    await Promise.all(
+      responsiblesArray.map(async (responsible: Responsible) => {
+        if (responsible.id) {
+          responsiblesToRemove = responsiblesToRemove?.filter(
+            (res) => res.id !== responsible.id,
+          );
+          const foundResponsible = this.responsibleRepository.findOne({
+            where: {
+              id: responsible.id,
+            },
+          });
+          const createdResponsible = this.responsibleRepository.create({
+            ...foundResponsible,
+            ...responsible,
+          });
+          await this.responsibleRepository.save(createdResponsible);
+        } else {
+          const createdResponsible = this.responsibleRepository.create({
+            ...responsible,
+            company: {
+              id: companyId,
+            },
+          });
+          await this.responsibleRepository.save(createdResponsible);
+        }
+      }),
+    );
+    await Promise.all(
+      responsiblesToRemove?.map(
+        async (oldResponsible) =>
+          oldResponsible.id &&
+          (await this.responsibleRepository.delete(oldResponsible.id)),
+      ),
+    );
   }
 
   async findAll(): Promise<Company[]> {
@@ -42,26 +99,32 @@ export class CompaniesService {
   }
 
   async findOne(id: string): Promise<Company> {
+    if (!id || checkValidUUID(id) === false) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundCompany: Company = await this.companyRepository.findOne({
       where: {
         id: id,
       },
-      relations: ['places', 'responsibles', 'user'],
+      relations: ['places', 'responsibles'],
     });
     if (!foundCompany) {
-      throw new Error('COMPANY_NOT_FOUND');
+      throw new NotFoundException('Empresa não encontrada');
     }
     return foundCompany;
   }
 
   async update(id: string, updateCompanyDto: UpdateCompanyDto) {
+    if (!id || checkValidUUID(id) === false) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundCompany: Company = await this.companyRepository.findOne({
       where: {
         id: id,
       },
     });
     if (!foundCompany) {
-      throw new Error('COMPANY_NOT_FOUND');
+      throw new NotFoundException('Empresa não encontrada');
     }
     const updatedCompany: Company = this.companyRepository.create({
       ...foundCompany,
@@ -70,17 +133,28 @@ export class CompaniesService {
     const savedCompany: Company = await this.companyRepository.save(
       updatedCompany,
     );
+
+    if (updateCompanyDto.responsibles) {
+      await this.handleUpdateCompanyResponsibles(
+        savedCompany.id,
+        updateCompanyDto.responsibles,
+        savedCompany.responsibles,
+      );
+    }
     return savedCompany;
   }
 
   async remove(id: string) {
+    if (!id || checkValidUUID(id) === false) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundCompany: Company = await this.companyRepository.findOne({
       where: {
         id: id,
       },
     });
     if (!foundCompany) {
-      throw new Error('COMPANY_NOT_FOUND');
+      throw new NotFoundException('Empresa não encontrada');
     }
     await this.companyRepository.delete(foundCompany.id);
   }
