+++ src/modules/users/users.service.ts
@@ -1,5 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { checkValidUUID } from 'src/common/checkValidUUID';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -12,6 +17,13 @@ export class UsersService {
     private userRepository: Repository<User>,
   ) {}
   async create(createUserDto: CreateUserDto) {
+    if (
+      !createUserDto.email ||
+      !createUserDto.name ||
+      !createUserDto.password
+    ) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const createdUser: User = this.userRepository.create(createUserDto);
 
     const savedUser = await this.userRepository.save(createdUser);
@@ -24,38 +36,47 @@ export class UsersService {
   }
 
   async findOne(id: string): Promise<User> {
+    if (!id || checkValidUUID(id) === false) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundUser: User = await this.userRepository.findOne({
       where: {
         id: id,
       },
-      relations: ['companies', 'ticket_to_attend', 'created_tickets'],
+      relations: ['ticket_to_attend', 'created_tickets'],
     });
     if (!foundUser) {
-      throw new Error('USER_NOT_FOUND');
+      throw new NotFoundException('Usuário não encontrado');
     }
     return foundUser;
   }
 
   async findByEmail(email: string): Promise<User> {
+    if (!email) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundUser: User = await this.userRepository.findOne({
       where: {
         email: email,
       },
     });
     if (!foundUser) {
-      throw new Error('USER_NOT_FOUND');
+      throw new NotFoundException('Usuário não encontrado');
     }
     return foundUser;
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+    if (!id || checkValidUUID(id) === false) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundUser: User = await this.userRepository.findOne({
       where: {
         id: id,
       },
     });
     if (!foundUser) {
-      throw new Error('USER_NOT_FOUND');
+      throw new NotFoundException('Usuário não encontrado');
     }
     const updatedUser: User = this.userRepository.create({
       ...foundUser,
@@ -66,13 +87,16 @@ export class UsersService {
   }
 
   async remove(id: string) {
+    if (!id || checkValidUUID(id) === false) {
+      throw new BadRequestException('Campos inválidos');
+    }
     const foundUser: User = await this.userRepository.findOne({
       where: {
         id: id,
       },
     });
     if (!foundUser) {
-      throw new Error('USER_NOT_FOUND');
+      throw new NotFoundException('Usuário não encontrado');
     }
     await this.userRepository.delete(foundUser.id);
   }
